Remove tiles from content view on double tap

+++ ScrollContent/ScrollContent/GameBoard.h
@@ -1,10 +1,10 @@
 #import <UIKit/UIKit.h>
 #import "Tile.h"
 
-extern int const kBoardWidth;
-extern int const kBoardHeight;
-extern int const kBoardTop;
-extern int const kBoardLeft;
+extern CGFloat const kBoardWidth;
+extern CGFloat const kBoardHeight;
+extern CGFloat const kBoardTop;
+extern CGFloat const kBoardLeft;
 
 @interface GameBoard: UIView
 + (CGPoint) snapToGrid:(CGPoint)pt;
